Move class into `glob` module.
