+++ Lib/pathlib/_abc.py
@@ -12,11 +12,10 @@ resemble pathlib's PurePath and Path respectively.
 """
 
 import functools
+import glob
 import operator
 from errno import ENOENT, ENOTDIR, EBADF, ELOOP, EINVAL
 from stat import S_ISDIR, S_ISLNK, S_ISREG, S_ISSOCK, S_ISBLK, S_ISCHR, S_ISFIFO
-
-from . import _glob
 
 #
 # Internals
@@ -43,12 +42,8 @@ def _ignore_error(exception):
 def _is_case_sensitive(parser):
     return parser.normcase('Aa') == 'Aa'
 
-#
-# Globbing helpers
-#
 
-
-class Globber(_glob.Globber):
+class Globber(glob._Globber):
     lstat = operator.methodcaller('lstat')
     scandir = operator.methodcaller('_scandir')
     add_slash = operator.methodcaller('joinpath', '')
@@ -699,7 +694,7 @@ class PathBase(PurePathBase):
             return iter([])
         if case_sensitive is None:
             case_sensitive = _is_case_sensitive(self.parser)
-        recursive = True if recurse_symlinks else _glob.no_recurse_symlinks
+        recursive = True if recurse_symlinks else glob._no_recurse_symlinks
         globber = self._globber(self.parser.sep, case_sensitive, recursive)
         return globber.selector(parts)
